src/requirements.txt
cachetools==4.1.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.1
google-api-core==1.22.0
google-auth==1.20.0
google-cloud-kms==1.4.0
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.30.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
protobuf==3.12.4
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycrypto==2.6.1
pytz==2020.1
requests==2.23.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.13
urllib3==1.25.9
uWSGI==2.0.19.1
Werkzeug==1.0.1
google.cloud.logging
